+++ src/constants.rs
@@ -15,7 +15,8 @@
 //! `LONG_DESCRIPTIVE_UPPER_CASE_NAMES`, but they can be brought into
 //! scope using a `let` binding:
 //!
-//! ```
+#![cfg_attr(feature = "basepoint-tables", doc = "```")]
+#![cfg_attr(not(feature = "basepoint-tables"), doc = "```ignore")]
 //! use curve25519_dalek::constants;
 //! use curve25519_dalek::traits::IsIdentity;
 //!
@@ -30,11 +31,13 @@
 
 use cfg_if::cfg_if;
 
-use crate::edwards::{CompressedEdwardsY, EdwardsBasepointTable};
+use crate::edwards::CompressedEdwardsY;
 use crate::montgomery::MontgomeryPoint;
-use crate::ristretto::CompressedRistretto;
-use crate::ristretto::RistrettoPoint;
+use crate::ristretto::{CompressedRistretto, RistrettoPoint};
 use crate::scalar::Scalar;
+
+#[cfg(feature = "basepoint-tables")]
+use crate::edwards::EdwardsBasepointTable;
 
 cfg_if! {
     if #[cfg(curve25519_dalek_backend = "fiat")] {
@@ -91,8 +94,11 @@ pub const BASEPOINT_ORDER: Scalar = Scalar {
     ],
 };
 
+#[cfg(feature = "basepoint-tables")]
 use crate::ristretto::RistrettoBasepointTable;
+
 /// The Ristretto basepoint, as a `RistrettoBasepointTable` for scalar multiplication.
+#[cfg(feature = "basepoint-tables")]
 pub static RISTRETTO_BASEPOINT_TABLE: &'static RistrettoBasepointTable = unsafe {
     // SAFETY: `RistrettoBasepointTable` is a `#[repr(transparent)]` newtype of
     // `EdwardsBasepointTable`