+++ CHANGELOG.md
@@ -25,6 +25,7 @@ major series.
 
 #### Other changes
 
+* Add `basepoint-tables` feature
 * Update Maintenance Policies for SemVer
 * Migrate documentation to docs.rs hosted
 * Fix backend documentation generation